Fix static plugin build
In some systems, Android being one of the most common examples,
plugins are usually built statically into the app (not much
of a plugin, though). There was a problem in the configure
script in which a custom GstShark variable was set
in the config.h rather that the standard GStreamer
required variable.